Handle 64 bit offsets.
No tests since llvm-mc takes 14 seconds on it. I will try to improve it
and then test.

Part of pr26208.

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@258129 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 7a02bc2182eb5883033fb57ebf58e8bbef103013)

apple-llvm-split-commit: e4e299a561ad7216c61290c502722f31838edd69
apple-llvm-split-dir: llvm/

+++ llvm/lib/MC/ELFObjectWriter.cpp
@@ -1273,7 +1273,7 @@ void ELFObjectWriter::writeObject(MCAssembler &Asm,
   uint64_t NaturalAlignment = is64Bit() ? 8 : 4;
   align(NaturalAlignment);
 
-  const unsigned SectionHeaderOffset = getStream().tell();
+  const uint64_t SectionHeaderOffset = getStream().tell();
 
   // ... then the section header table ...
   writeSectionHeader(Layout, SectionIndexMap, SectionOffsets);
